requirements.txt
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Cython==0.29.14
dyNET==2.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
Flask-Mail==0.9.1
flask-swagger-ui==3.36.0
gensim==3.8.3
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
JPype1==1.2.1
kiwisolver==1.3.1
linecache2==1.0.0
log4p==2019.7.13.3
MarkupSafe==1.1.1
matplotlib==3.4.1
mockupdb==1.8.0
nltk==3.6.2
numpy==1.20.2
pandas==1.2.4
passlib==1.7.4
Pillow==8.2.0
protobuf==3.15.8
psutil==5.8.0
pycairo==1.20.0
pyjnius==1.3.0
PyJWT==1.7.1
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.0
python-igraph==0.9.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
scikit-learn==0.23.2
scipy==1.6.3
six==1.15.0
smart-open==5.0.0
stanza==1.2
tabulate==0.8.9
texttable==1.6.3
threadpoolctl==2.1.0
torch==1.8.1
tqdm==4.60.0
traceback2==1.4.0
typing-extensions==3.7.4.3
unittest2==1.1.0
urllib3==1.26.4
Werkzeug==1.0.1
